simplified the configuration code
