+++ examples/basic.rs
@@ -43,23 +43,9 @@ fn setup(
     mut quadtrees: ResMut<TerrainViewComponents<Quadtree>>,
     mut view_configs: ResMut<TerrainViewComponents<TerrainViewConfig>>,
 ) {
-    let mut preprocessor = Preprocessor::default();
-    let mut loader = AttachmentFromDiskLoader::default();
-
-    // Configure all the important properties of the terrain, as well as its attachments.
-    let mut config = TerrainConfig::new(
+    let mut loader = AttachmentFromDiskLoader::new(LOD_COUNT, PATH.to_string());
+    loader.add_base_attachment(
         &plugin_config,
-        TERRAIN_SIZE,
-        LOD_COUNT,
-        HEIGHT,
-        NODE_ATLAS_SIZE,
-        PATH.to_string(),
-    );
-
-    config.add_base_attachment_from_disk(
-        &plugin_config,
-        &mut preprocessor,
-        &mut loader,
         TileConfig {
             path: "assets/terrain/source/height".to_string(),
             size: TERRAIN_SIZE,
@@ -69,18 +55,16 @@ fn setup(
 
     // Preprocesses the terrain data.
     // Todo: Should be commented out after the first run.
-    // preprocessor.preprocess(&config);
+    // loader.preprocess();
 
-    load_node_config(&mut config);
-
-    // Create the terrain.
-    let terrain = commands
-        .spawn((
-            TerrainBundle::new(config.clone()),
-            loader,
-            materials.add(TerrainMaterial {}),
-        ))
-        .id();
+    // Configure all the important properties of the terrain, as well as its attachments.
+    let config = plugin_config.configure_terrain(
+        TERRAIN_SIZE,
+        LOD_COUNT,
+        HEIGHT,
+        NODE_ATLAS_SIZE,
+        PATH.to_string(),
+    );
 
     // Configure the quality settings of the terrain view. Adapt the settings to your liking.
     let view_config = TerrainViewConfig {
@@ -91,6 +75,15 @@ fn setup(
         view_distance: 4.0,
         ..default()
     };
+
+    // Create the terrain.
+    let terrain = commands
+        .spawn((
+            TerrainBundle::new(config.clone()),
+            loader,
+            materials.add(TerrainMaterial {}),
+        ))
+        .id();
 
     // Create the view.
     let view = commands
